requirements.txt
gradio>=4.0.0
groq>=0.4.0
python-dotenv>=1.0.0